refactor: Move ngram decoder to script
